+++ admin/section-controls.php
@@ -38,14 +38,14 @@ global $post;
 		</ul>
 	</div>
 
-	<a href="#" class="slide-toggle-element slide-toggle-meta-dropdown mesh-more-section-options" data-toggle=".mesh-section-meta-dropdown-<?php esc_attr_e( $section->ID ); ?>"><?php _e( 'More Options' ); ?></a>
+	<a href="#" class="slide-toggle-element slide-toggle-meta-dropdown mesh-more-section-options" data-toggle=".mesh-section-meta-dropdown-<?php esc_attr_e( $section->ID ); ?>"><?php _e( 'More Options', 'mesh' ); ?></a>
 </div>
 
-<div class="mesh-section-meta-dropdown mesh-section-meta-dropdown-<?php esc_attr_e( $section->ID ); ?> mesh-row hide">
+<div class="mesh-section-meta-dropdown mesh-section-meta-dropdown-<?php echo esc_attr( $section->ID ); ?> mesh-row hide">
 	<div class="mesh-columns-9 mesh-table">
 		<div class="mesh-row">
-			<label for="mesh-sections[<?php esc_attr_e( $section->ID ); ?>][css-class]">
-				<?php esc_html_e( 'CSS Class', 'mesh' ); ?> <input type="text" name="mesh-sections[<?php esc_attr_e( $section->ID ); ?>][css_class]" class="mesh-section-class" value="<?php esc_attr_e( $css_class ); ?>" />
+			<label for="mesh-sections[<?php echo esc_attr( $section->ID ); ?>][css-class]">
+				<?php esc_html_e( 'CSS Class', 'mesh' ); ?> <input type="text" name="mesh-sections[<?php echo esc_attr( $section->ID ); ?>][css_class]" class="mesh-section-class" value="<?php echo esc_attr( $css_class ); ?>" />
 			</label>
 		</div>
 
@@ -53,23 +53,23 @@ global $post;
 			<?php if ( ! has_term( 'reference', 'mesh_template_types', $post ) ) : ?>
 			<ul class="inline-block-list space-left">
 				<li>
-					<label for="mesh-section[<?php esc_attr_e( $section->ID ); ?>][collapse]">
-						<?php esc_html_e( 'Collapse Padding', 'mesh' ); ?> <input type="checkbox" name="mesh-sections[<?php esc_attr_e( $section->ID ); ?>][collapse]" class="mesh-section-collapse-input" value="1" <?php if ( get_post_meta( $section->ID, '_mesh_collapse', true ) ) : ?>checked<?php endif; ?> />
+					<label for="mesh-section[<?php echo esc_attr( $section->ID ); ?>][collapse]">
+						<?php esc_html_e( 'Collapse Padding', 'mesh' ); ?> <input type="checkbox" name="mesh-sections[<?php echo esc_attr( $section->ID ); ?>][collapse]" class="mesh-section-collapse-input" value="1" <?php if ( get_post_meta( $section->ID, '_mesh_collapse', true ) ) : ?>checked<?php endif; ?> />
 					</label>
 				</li>
 
 				<?php if ( 2 === count( $blocks ) ) : ?>
 					<li>
-						<label for="mesh-section[<?php esc_attr_e( $section->ID ); ?>][push-pull]">
-							<?php esc_html_e( 'Push/Pull', 'mesh' ); ?> <input type="checkbox" name="mesh-sections[<?php esc_attr_e( $section->ID ); ?>][push_pull]" class="mesh-section-push" value="1" <?php if ( get_post_meta( $section->ID, '_mesh_push_pull', true ) ) : ?>checked<?php endif; ?> />
+						<label for="mesh-section[<?php echo esc_attr( $section->ID ); ?>][push-pull]">
+							<?php esc_html_e( 'Push/Pull', 'mesh' ); ?> <input type="checkbox" name="mesh-sections[<?php echo esc_attr( $section->ID ); ?>][push_pull]" class="mesh-section-push" value="1" <?php if ( get_post_meta( $section->ID, '_mesh_push_pull', true ) ) : ?>checked<?php endif; ?> />
 						</label>
 					</li>
 				<?php endif; ?>
 
 				<?php if ( 1 < count( $blocks ) ) : ?>
 				<li>
-					<label for="mesh-sections[<?php esc_attr_e( $section->ID ); ?>][lp-equal]">
-						<?php esc_html_e( 'Equalize', 'mesh' ); ?> <input type="checkbox" name="mesh-sections[<?php esc_attr_e( $section->ID ); ?>][lp_equal]" class="mesh-section-equalize" value="1" <?php if ( get_post_meta( $section->ID, '_mesh_lp_equal', true ) ) : ?>checked<?php endif; ?> />
+					<label for="mesh-sections[<?php echo esc_attr( $section->ID ); ?>][lp-equal]">
+						<?php esc_html_e( 'Equalize', 'mesh' ); ?> <input type="checkbox" name="mesh-sections[<?php echo esc_attr( $section->ID ); ?>][lp_equal]" class="mesh-section-equalize" value="1" <?php if ( get_post_meta( $section->ID, '_mesh_lp_equal', true ) ) : ?>checked<?php endif; ?> />
 					</label>
 				</li>
 				<?php endif; ?>
@@ -84,8 +84,8 @@ global $post;
 				<a href="#" class="mesh-featured-image-choose button"><?php esc_html_e( 'Set Background Image', 'mesh' ); ?></a>
 			<?php else : ?>
 			<?php $featured_image = wp_get_attachment_image_src( $featured_image_id, array( 160, 60 ) ); ?>
-				<a href="#" class="mesh-featured-image-choose right" data-mesh-section-featured-image="<?php esc_attr_e( $featured_image_id ); ?>"><img src="<?php echo esc_url( $featured_image[0] ); ?>" /></a>
-				<a href="#" class="mesh-featured-image-trash dashicons-before dashicons-dismiss" data-mesh-section-featured-image="<?php esc_attr_e( $featured_image_id ); ?>"></a>
+				<a href="#" class="mesh-featured-image-choose right" data-mesh-section-featured-image="<?php echo esc_attr( $featured_image_id ); ?>"><img src="<?php echo esc_url( $featured_image[0] ); ?>" /></a>
+				<a href="#" class="mesh-featured-image-trash dashicons-before dashicons-dismiss" data-mesh-section-featured-image="<?php echo esc_attr( $featured_image_id ); ?>"></a>
 			<?php endif; ?>
 		</div>
 	</div>